+++ game/gui/widgets/gui_entities.cpp
@@ -8,14 +8,20 @@
 #include "r_window.h"
 
 #include "log.h"
+#include "scene_list.h"
+#include "s_collider.h"
+#include "serialization.h"
 
 using namespace Verse;
 
 namespace {
     ui32 n = 0;
     std::map<str, std::function<void(Config&, EntityID)>> c_funcs;
+
     Component::Camera* prev_cam;
     Component::Camera* tile_cam;
+
+    int scene_index = -1;
 }
 
 void components(Config &c, Signature mask, EntityID e);
@@ -25,6 +31,12 @@ void Gui::entities(Config &c) {
     
     ImGui::Text("scene: %s", ((c.active_scene != nullptr) ? c.active_scene->name.c_str() : "NULL"));
     ImGui::Text("number of entities: %d", n);
+    
+    if (ImGui::SmallButton("save (temp)"))
+        Serialization::saveScene(c.active_scene, c);
+    ImGui::SameLine();
+    if (ImGui::SmallButton("save (project)"))
+        Serialization::saveScene(c.active_scene, c, true);
     
     static ImGuiTableFlags flags = ImGuiTableFlags_PadOuterX | ImGuiTableFlags_RowBg;
     
@@ -96,7 +108,7 @@ void c_collider(Config &c, EntityID e) {
     ImGui::TableSetColumnIndex(1);
     str layer_label = "##collayer" + std::to_string(e);
     ImGui::SetNextItemWidth(ImGui::GetColumnWidth());
-    ImGui::Combo(layer_label.c_str(), col->layer, Component::c_layers_name);
+    ImGui::Combo(layer_label.c_str(), col->layer, System::Collider::layers_name);
 }
 
 void c_texture(Config &c, EntityID e) {
@@ -106,7 +118,6 @@ void c_texture(Config &c, EntityID e) {
     ImGui::Text("res");
     
     ImGui::TableSetColumnIndex(1);
-    
     float line_height = ImGui::GetStyle().FramePadding.y * 2.0f + ImGui::CalcTextSize("X").y;
     ImVec2 button_size = { line_height + 3.0f, line_height };
     if (ImGui::Button("L", button_size))
@@ -158,18 +169,20 @@ void c_tilemap(Config &c, EntityID e) {
     }
     
     ImGui::TableSetColumnIndex(0);
-    if (c.tilemap_editor and c.current_tilemap_edit == tile) {
-        if (ImGui::Button("save")) {
-            c.tilemap_editor = false;
+    if (c.tme_active and c.tme_curr_tmap == tile) {
+        if (ImGui::SmallButton("save")) {
+            c.tme_active = false;
             c.use_light = true;
-            c.current_tilemap_edit = nullptr;
+            c.tme_curr_tmap = nullptr;
+            c.tme_curr_id = 0;
             c.active_camera = prev_cam;
         }
     } else {
-        if (ImGui::Button("edit")) {
-            c.tilemap_editor = true;
+        if (ImGui::SmallButton("edit")) {
+            c.tme_active = true;
             c.use_light = false;
-            c.current_tilemap_edit = tile;
+            c.tme_curr_tmap = tile;
+            c.tme_curr_id = e;
             prev_cam = c.active_camera;
             tile_cam->pos = prev_cam->pos;
             c.active_camera = tile_cam;
@@ -196,8 +209,23 @@ void c_actor(Config &c, EntityID e) {
     ImGui::TableNextRow();
     
     Verse::Gui::draw_bool(actor->has_gravity, "gravity", e);
+    ImGui::TableNextRow();
     
-    //TODO: Controller and collision mask
+    //TODO: Controller
+    
+    ImGui::TableSetColumnIndex(0);
+    ImGui::Text("collision mask");
+    ImGui::TableSetColumnIndex(1);
+    int i = 0;
+    for (str layer : System::Collider::layers_name) {
+        str layer_label = layer + "##" + std::to_string(e);
+        bool layer_active = actor->collision_mask[i];
+        
+        ImGui::SetNextItemWidth(ImGui::GetColumnWidth());
+        if (ImGui::Selectable(layer_label.c_str(), layer_active))
+            (actor->collision_mask[i]) ? actor->collision_mask.reset(i) : actor->collision_mask.set(i);
+        i++;
+    }
 }
 
 void c_light(Config &c, EntityID e) {
@@ -214,11 +242,67 @@ void c_camera(Config &c, EntityID e) {
 }
 
 void c_fire(Config &c, EntityID e) {
-    //Component::Fire* fire = c.active_scene->getComponent<Component::Fire>(e);
+    Component::Fire* fire = c.active_scene->getComponent<Component::Fire>(e);
+    
+    Verse::Gui::draw_vec2(fire->transform.x, fire->transform.y, "pos", e);
+    ImGui::TableNextRow();
+    Verse::Gui::draw_vec2(fire->transform.w, fire->transform.h, "size", e); //TODO: Change fire data size
+    ImGui::TableNextRow();
+    Verse::Gui::draw_vec2(fire->offset.x, fire->offset.y, "offset", e);
+    ImGui::TableNextRow();
+    
+    Verse::Gui::draw_float(fire->freq, "freq", e);
+    ImGui::TableNextRow();
+    Verse::Gui::draw_ui8(fire->octaves, "octaves", e);
+    ImGui::TableNextRow();
+    Verse::Gui::draw_vec2(fire->dir.x, fire->dir.y, "dir", e);
+    ImGui::TableNextRow();
+    
+    Verse::Gui::draw_int(fire->layer, "layer", e);
+    ImGui::TableNextRow();
+    Verse::Gui::draw_ui8(fire->fps, "fps", e);
+    ImGui::TableNextRow();
+    
+    ImGui::TableSetColumnIndex(0);
+    ImGui::Text("texture");
+    
+    ImGui::TableSetColumnIndex(1);
+    float line_height = ImGui::GetStyle().FramePadding.y * 2.0f + ImGui::CalcTextSize("X").y;
+    ImVec2 button_size = { line_height + 3.0f, line_height };
+    if (ImGui::Button("L", button_size))
+        Graphics::Texture::loadTexture(fire->flame_tex_res, fire->flame_tex);
+    
+    ImGui::SameLine();
+    str res_label = "##res" + std::to_string(e);
+    ImGui::SetNextItemWidth(ImGui::GetColumnWidth());
+    ImGui::InputText(res_label.c_str(), &fire->flame_tex_res);
 }
 
 void c_scene_transition(Config &c, EntityID e) {
-    //Component::SceneTransition* trans = c.active_scene->getComponent<Component::SceneTransition>(e);
+    Component::SceneTransition* trans = c.active_scene->getComponent<Component::SceneTransition>(e);
+    
+    if (scene_index == -1) {
+        auto it = std::find(scenes.begin(), scenes.end(), trans->scene_name);
+        if (it != scenes.end())
+            scene_index = 0;
+        else
+            scene_index = (int)std::distance(scenes.begin(), it);
+    }
+    
+    ImGui::TableSetColumnIndex(0);
+    ImGui::Text("new scene");
+    
+    ImGui::TableSetColumnIndex(1);
+    str layer_label = "##scenetrans" + std::to_string(e);
+    ImGui::SetNextItemWidth(ImGui::GetColumnWidth());
+    if (ImGui::Combo(layer_label.c_str(), &scene_index, scenes) and trans->scene_name != scenes[scene_index]) {
+        trans->scene_name = scenes[scene_index];
+        trans->to_scene = new Scene();
+        Serialization::loadScene(trans->scene_name, trans->to_scene, c);
+    }
+    ImGui::TableNextRow();
+    
+    Verse::Gui::draw_vec2(trans->to_pos.x, trans->to_pos.y, "new scene pos", e);
 }
 
 void c_player(Config &c, EntityID e) {
@@ -235,7 +319,7 @@ void components(Config &c, Signature mask, EntityID e) {
         c_funcs["light"] = c_light;
         c_funcs["camera"] = c_camera;
         c_funcs["fire"] = c_fire;
-        c_funcs["sceneTransition"] = c_collider;
+        c_funcs["sceneTransition"] = c_scene_transition;
         c_funcs["player"] = c_player;
     }