+++ game/systems/s_actor.cpp
@@ -7,13 +7,12 @@
 #include "game.h"
 
 #include "s_collider.h"
+#include "s_scene_transition.h"
+#include "s_fire.h"
+#include "r_textures.h"
 
 using namespace Verse;
 using namespace State;
-
-namespace {
-    std::bitset<MAX_COLLISION_LAYERS> collision_layers;
-}
 
 void System::Actor::update(Config &c) {
     for (EntityID e : SceneView<Component::Actor>(*c.active_scene)) {
@@ -52,40 +51,32 @@ bool System::Actor::move(Config &c, EntityID eid, State::StateType state) {
     while (to_move.x != 0) {
         collider->transform += Vec2::i * sx;
         
-        collision_layers = System::Collider::checkCollisions(c, eid);
+        ui8 colliding = collisions(c, eid);
         
-        if (collision_layers[Component::ColliderLayers::SCENE])
+        if (colliding == Colliding::Exit)
             return false;
         
-        if (collision_layers[Component::ColliderLayers::WATER]) {
-            actor->damage();
-            return false;
-        }
-        
-        if (collision_layers.any()) {
+        if (colliding == Colliding::Solid) {
             collider->transform -= Vec2::i * sx;
             
             to_move.x = 0;
             actor->remainder.x = 0;
             actor->vel.x = 0;
-        } else {
+        }
+        
+        if (colliding == Colliding::Transparent) {
             to_move.x -= sx;
             
             //Check on ground
             if (to_move.y == 0) {
                 collider->transform += Vec2::j;
                 
-                collision_layers = System::Collider::checkCollisions(c, eid);
+                ui8 colliding = collisions(c, eid);
                 
-                if (collision_layers[Component::ColliderLayers::SCENE])
+                if (colliding == Colliding::Exit)
                     return false;
                 
-                if (collision_layers[Component::ColliderLayers::WATER]) {
-                    actor->damage();
-                    return false;
-                }
-                
-                if (collision_layers.none() and p_state != nullptr)
+                if (colliding != Colliding::Solid and p_state != nullptr)
                     p_state->jump.handle(Player::FallEvent());
                 
                 collider->transform -= Vec2::j;
@@ -98,17 +89,12 @@ bool System::Actor::move(Config &c, EntityID eid, State::StateType state) {
     while (to_move.y != 0) {
         collider->transform += Vec2::j * sy;
         
-        collision_layers = System::Collider::checkCollisions(c, eid);
+        ui8 colliding = collisions(c, eid);
         
-        if (collision_layers[Component::ColliderLayers::SCENE])
+        if (colliding == Colliding::Exit)
             return false;
         
-        if (collision_layers[Component::ColliderLayers::WATER]) {
-            actor->damage();
-            return false;
-        }
-        
-        if (collision_layers.any()) {
+        if (colliding == Colliding::Solid) {
             collider->transform -= Vec2::j * sy;
             
             to_move.y = 0;
@@ -117,7 +103,9 @@ bool System::Actor::move(Config &c, EntityID eid, State::StateType state) {
             
             if (sy > 0 and p_state != nullptr)
                 p_state->jump.handle(Player::TouchGroundEvent());
-        } else {
+        }
+        
+        if (colliding == Colliding::Transparent) {
             to_move.y -= sy;
             
             p_state->jump.handle(Player::FallEvent());
@@ -135,4 +123,58 @@ bool System::Actor::move(Config &c, EntityID eid, State::StateType state) {
 #endif
         
     return true;
+}
+
+ui8 System::Actor::collisions(Config &c, EntityID eid) {
+    System::Collider::CollisionInfo collision_info = System::Collider::checkCollisions(c, eid);
+    
+    bool solid = false;
+    
+    for (System::Collider::CollisionInfoPair collision : collision_info) {
+        
+        if (collision.second[System::Collider::Layers::Scene]) {
+            Component::SceneTransition* c_transition = c.active_scene->getComponent<Component::SceneTransition>(collision.first);
+            if (c_transition != nullptr) {
+                System::SceneTransition::handle(c, c_transition);
+                return Colliding::Exit;
+            }
+        }
+        
+        if (collision.second[System::Collider::Layers::Water]) {
+            Component::Actor* actor = c.active_scene->getComponent<Component::Actor>(eid);
+            actor->damage();
+            return Colliding::Exit;
+        }
+        
+        if (collision.second[System::Collider::Layers::Checkpoint]) {
+            Component::Collider* c_col = c.active_scene->getComponent<Component::Collider>(collision.first);
+            Component::Fire* fire = c.active_scene->getComponent<Component::Fire>(collision.first);
+            Component::Light* light = c.active_scene->getComponent<Component::Light>(collision.first);
+            
+            if (fire == nullptr) {
+                c.active_scene->checkpoints.push_back(c_col->transform.pos() - Vec2(0,4));
+                
+                fire = c.active_scene->addComponent<Component::Fire>(collision.first);
+                Vec2 offset = Vec2(c_col->transform.w * 0.5f - 5.5, -c_col->transform.h);
+                fire->transform = Rect2(c_col->transform.pos() + offset, Vec2(11, 11));
+                fire->dir = Vec2::j;
+                fire->fps = 3;
+                fire->freq = 16;
+                fire->octaves = 16;
+                fire->seed = (rand() / RAND_MAX) % 1000;
+                fire->layer = 0;
+                Graphics::Texture::loadTexture("res/graphics/flame.png", fire->flame_tex);
+                System::Fire::init(fire);
+                
+                light = c.active_scene->addComponent<Component::Light>(collision.first);
+                light->pos = Vec2(c_col->transform.w * 0.5f, -c_col->transform.h * 0.5f);
+                light->radius = 50;
+            }
+        }
+        
+        if (collision.second[System::Collider::Layers::Ground])
+            solid = true;
+    }
+    
+    return solid ? Colliding::Solid : Colliding::Transparent;
 }