+++ game/systems/s_collider.cpp
@@ -5,10 +5,7 @@
 #include "s_collider.h"
 
 #include "s_tilemap.h"
-#include "s_scene_transition.h"
-#include "s_fire.h"
 #include "r_renderer.h"
-#include "r_textures.h"
 
 #include "log.h"
 
@@ -54,13 +51,17 @@ std::vector<EntityID> System::Collider::checkObjectCollisions(Config &c, EntityI
     return collisions;
 }
 
-bool System::Collider::checkTilemapCollision(Component::Collider* test_col, Component::Collider* tile_col, Component::Tilemap* tilemap) {
-    bool is_colliding_with_tile = false;
+System::Collider::CollisionInfoPair System::Collider::checkTilemapCollision(Config &c, EntityID eid, Component::Collider* test_col) {
+    Component::Collider* tile_col = c.active_scene->getComponent<Component::Collider>(eid);
+    Component::Tilemap* tilemap = c.active_scene->getComponent<Component::Tilemap>(eid);
     
     Vec2 rel_pos = test_col->transform.pos() - tile_col->transform.pos();
     Vec2 upleft = Vec2(floor(rel_pos.x / tilemap->tex_size.x), floor(rel_pos.y / tilemap->tex_size.y));
     Vec2 downright = Vec2(floor((rel_pos.x + test_col->transform.w - 1) / tilemap->tex_size.x),
                           floor((rel_pos.y + test_col->transform.h - 1) / tilemap->tex_size.y));
+    
+    System::Collider::CollisionInfoPair tile_collisions;
+    tile_collisions.first = eid;
     
     for (int i = upleft.x; i <= downright.x; i++) {
         if (i < 0 or i >= tilemap->tiles[0].size())
@@ -69,15 +70,18 @@ bool System::Collider::checkTilemapCollision(Component::Collider* test_col, Comp
             if (j < 0 or j >= tilemap->tiles.size() - 1)
                 continue;
             if (tilemap->tiles[j][i] < 255)
-                is_colliding_with_tile = true;
+                tile_collisions.second.set(System::Collider::Layers::Ground);
         }
     }
     
-    return is_colliding_with_tile;
+    if (tile_collisions.second.none())
+        test_col->is_colliding = false;
+    
+    return tile_collisions;
 }
 
-std::bitset<MAX_COLLISION_LAYERS> System::Collider::checkCollisions(Config &c, EntityID eid) {
-    std::bitset<MAX_COLLISION_LAYERS> collision_layers;
+System::Collider::CollisionInfo System::Collider::checkCollisions(Config &c, EntityID eid) {
+    System::Collider::CollisionInfo col_info;
     
     Component::Collider* collider = c.active_scene->getComponent<Component::Collider>(eid);
     std::vector<EntityID> collisions = System::Collider::checkObjectCollisions(c, eid);
@@ -86,62 +90,17 @@ std::bitset<MAX_COLLISION_LAYERS> System::Collider::checkCollisions(Config &c, E
         for (const EntityID &e : collisions) {
             Component::Collider* c_col = c.active_scene->getComponent<Component::Collider>(e);
             
-            //Tilemap
             Component::Tilemap* c_tile = c.active_scene->getComponent<Component::Tilemap>(e);
             if (c_tile != nullptr) {
-                bool is_colliding_with_tile = System::Collider::checkTilemapCollision(collider, c_col, c_tile);
-                if (is_colliding_with_tile)
-                    collision_layers.set(Component::ColliderLayers::GROUND);
-                else
-                    collider->is_colliding = false;
+                col_info[e] = System::Collider::checkTilemapCollision(c, e, collider).second;
                 continue;
             }
             
-            collision_layers.set(c_col->layer);
-            
-            //Scene Transition
-            if (c_col->layer == Component::ColliderLayers::SCENE) {
-                Component::SceneTransition* c_transition = c.active_scene->getComponent<Component::SceneTransition>(e);
-                System::SceneTransition::handle(c, c_transition);
-                continue;
-            }
-            
-            //Checkpoint
-            if (c_col->layer == Component::ColliderLayers::CHECKPOINT) {
-                //TODO: MOVE TO EXTERNAL FUNCTION
-                auto it = std::find(c.active_scene->checkpoints.begin(), c.active_scene->checkpoints.end(), c_col->transform.pos());
-                if (it == c.active_scene->checkpoints.end())
-                    c.active_scene->checkpoints.push_back(c_col->transform.pos() - collider->transform.size() * 0.5f);
-                
-                Component::Fire* fire = c.active_scene->getComponent<Component::Fire>(e);
-                if (fire == nullptr) {
-                    fire = c.active_scene->addComponent<Component::Fire>(e);
-                    Vec2 offset = Vec2(c_col->transform.w * 0.5f - 5.5, -c_col->transform.h);
-                    fire->transform = Rect2(c_col->transform.pos() + offset, Vec2(11, 11));
-                    fire->dir = Vec2::j;
-                    fire->fps = 3;
-                    fire->freq = 16;
-                    fire->octaves = 16;
-                    fire->seed = (rand() / RAND_MAX) % 1000;
-                    fire->layer = 0;
-                    Graphics::Texture::loadTexture("res/graphics/flame.png", fire->flame_tex);
-                    System::Fire::init(fire);
-                }
-                
-                Component::Light* light = c.active_scene->getComponent<Component::Light>(e);
-                if (light == nullptr) {
-                    light = c.active_scene->addComponent<Component::Light>(e);
-                    light->pos = Vec2(c_col->transform.w * 0.5f, -c_col->transform.h * 0.5f);
-                    light->radius = 50;
-                }
-                
-                collision_layers.reset(Component::ColliderLayers::CHECKPOINT);
-                continue;
-            }
+            col_info[e].set(c_col->layer);
         }
     }
     
-    return collision_layers;
+    return col_info;
 }
 
 
@@ -155,3 +114,21 @@ void System::Collider::render(Config &c) {
         Graphics::Renderer::renderDebugCollider(c, collider->transform, collider->is_colliding);
     }
 }
+
+/*if (tiles.w == 0)
+    tiles = Rect2(i * tilemap->tex_size.x, j * tilemap->tex_size.y, tilemap->tex_size.x, tilemap->tex_size.y);
+
+if (i * tilemap->tex_size.x < tiles.x) {
+    tiles.w += tiles.x - i * tilemap->tex_size.x;
+    tiles.x = i * tilemap->tex_size.x;
+}
+if (j * tilemap->tex_size.y < tiles.y) {
+    tiles.h += tiles.y - j * tilemap->tex_size.y;
+    tiles.y = j * tilemap->tex_size.y;
+}
+if (i * tilemap->tex_size.x >= tiles.x + tilemap->tex_size.x) {
+    tiles.w += i * tilemap->tex_size.x - tiles.x;
+}
+if (j * tilemap->tex_size.y >= tiles.y + tilemap->tex_size.y) {
+    tiles.h += j * tilemap->tex_size.y - tiles.y;
+}*/