+++ game/systems/s_collider.h
@@ -5,11 +5,35 @@
 #pragma once
 
 #include "config.h"
+#include <map>
 
 namespace Verse::System::Collider
 {
+    using CollisionInfo = std::map<EntityID, std::bitset<MAX_COLLISION_LAYERS>>;
+    using CollisionInfoPair = std::pair<EntityID, std::bitset<MAX_COLLISION_LAYERS>>;
+
+    enum Layers {
+        Ground,
+        Actor,
+        Event,
+        Scene,
+        Water,
+        Checkpoint,
+        Platform,
+    };
+
+    static std::vector<str> layers_name = {
+        "ground",
+        "actor",
+        "event",
+        "scene",
+        "water",
+        "checkpoint",
+        "platform",
+    };
+
     std::vector<EntityID> checkObjectCollisions(Config &c, EntityID eid);
-    bool checkTilemapCollision(Component::Collider* test_col, Component::Collider* tile_col, Component::Tilemap* tilemap);
-    std::bitset<MAX_COLLISION_LAYERS> checkCollisions(Config &c, EntityID eid);
+    CollisionInfoPair checkTilemapCollision(Config &c, EntityID eid, Component::Collider* test_col);
+    CollisionInfo checkCollisions(Config &c, EntityID eid);
     void render(Config &c);
 }