+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Render collision boxes
 - Window to scene coordinates
 - Tilemap editor
+- Save scenes
+- Collider overhaul
 #### Changed
 - Light now cleans every frame